+++ orangecontrib/text/widgets/tests/__init__.py
@@ -1,0 +1,2 @@
+from Orange.data import Table
+Table.LOCKING = False